+++ src/test/testingSchemas.ts
@@ -22,12 +22,15 @@ import makeRemoteExecutableSchema, {
 import introspectSchema from '../stitching/introspectSchema';
 import { PubSub } from 'graphql-subscriptions';
 
+export type Location = {
+  name: string;
+  coordinates: string;
+};
+
 export type Property = {
   id: string;
   name: string;
-  location: {
-    name: string;
-  };
+  location: Location;
 };
 
 export type Product = {
@@ -84,6 +87,7 @@ export const sampleData: {
       name: 'Super great hotel',
       location: {
         name: 'Helsinki',
+        coordinates: '60.1698° N, 24.9383° E'
       },
     },
     p2: {
@@ -91,6 +95,7 @@ export const sampleData: {
       name: 'Another great hotel',
       location: {
         name: 'San Francisco',
+        coordinates: '37.7749° N, 122.4194° W'
       },
     },
     p3: {
@@ -98,6 +103,7 @@ export const sampleData: {
       name: 'BedBugs - The Affordable Hostel',
       location: {
         name: 'Helsinki',
+        coordinates: '60.1699° N, 24.9384° E'
       },
     },
   },
@@ -629,6 +635,7 @@ const bookingResolvers: IResolvers = {
 const subscriptionTypeDefs = `
   type Notification{
     text: String
+    throwError: String
   }
 
   type Query{
@@ -653,6 +660,11 @@ const subscriptionResolvers: IResolvers = {
         subscriptionPubSub.asyncIterator(subscriptionPubSubTrigger),
     },
   },
+  Notification: {
+    throwError: () => {
+      throw new Error('subscription field error');
+    }
+  }
 };
 
 export const propertySchema: GraphQLSchema = makeExecutableSchema({